Update models foreign key Update emloyee controllers

+++ migrations/20201211023234-create-time-off.js
@@ -30,7 +30,7 @@ module.exports = {
         onDelete: 'SET NULL'
       },
       day_off: {
-        type: Sequelize.DOUBLE
+        type: Sequelize.DECIMAL
       },
       createdAt: {
         allowNull: false,
